FORTA NETWORK - Token App & Contract Position Integration
Sorry beforehand for my ignorance and mistakes, it's the first time I commit a PR.
I'm trying to integrate FORT token in polygon as an APP token. And delegated FORT staking positions as an CONTRACT position.
Need just a little extra help or push in deploying the app token and contract position.
https://docs.forta.network/en/latest/smart-contracts/